pyproject.toml
[project]
name = "pfsconfig_redaction"
version = "0.1.0"
description = "Redact information in pfsConfig (PFSF) files for raw data deliverty to PIs."
authors = [{ name = "Masato Onodera", email = "monodera@naoj.org" }]
dependencies = [
    "pfs-datamodel @ git+https://github.com/Subaru-PFS/datamodel.git",
    "astropy>=5.2",                                                    # required by pfs-datamodel
    "numpy>=1.24.4",                                                   # required by pfs-datamodel
    "scipy>=1.9.3",                                                    # required by pfs-datamodel
]
requires-python = ">=3.9,<3.13"
readme = "README.md"
license = { text = "MIT" }

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
